requirements.txt
asgiref==3.7.2
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
Django==4.2.5
django-cors-headers==4.3.0
django-environ==0.11.2
django-extensions==3.2.3
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
environ==1.0
exceptiongroup @ file:///croot/exceptiongroup_1668714342571/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
ipython @ file:///croot/ipython_1694181358621/work
jedi @ file:///tmp/build/80754af9/jedi_1644315229345/work
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
phonenumbers==8.13.22
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==10.0.1
postgres==4.0
prompt-toolkit @ file:///croot/prompt-toolkit_1672387306916/work
psycopg2-binary==2.9.9
psycopg2-pool==1.1
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
Pygments @ file:///croot/pygments_1684279966437/work
PyJWT==2.8.0
pytz==2023.3.post1
six @ file:///tmp/build/80754af9/six_1644875935023/work
sqlparse==0.4.4
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
traitlets @ file:///croot/traitlets_1671143879854/work
typing_extensions==4.8.0
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
